requirements.txt
elevate==0.1.3
peid==2.2.1
pydantic==2.9.2
PyYAML==6.0.2

pylint==3.3.1
mypy==1.11.2